Refactored IGameService Part 1
Game classes now more split logic, with abstract Game class
