+++ src/cards/Data/Enums/Card/Suit.cs
@@ -1,0 +1,10 @@
+﻿namespace cards.Data.Enums.Card
+{
+    public enum Suit
+    {
+        Clovers,
+        Peaks,
+        Hearts,
+        Diamonds
+    }
+}
