+++ src/cards/Data/Enums/Card/Value.cs
@@ -1,0 +1,19 @@
+﻿namespace cards.Data.Enums.Card
+{
+    public enum Value
+    {
+        Two,
+        Three,
+        Four,
+        Five,
+        Six,
+        Seven,
+        Eight,
+        Nine,
+        Ten,
+        Jack,
+        Queen,
+        King,
+        Ace
+    }
+}
