+++ src/cards/Data/Game/Game.cs
@@ -2,47 +2,35 @@ using cards.Data.Game.Implementations;
 
 namespace cards.Data.Game;
 
-public interface IGameService
+public abstract class Game
 {
-    public static string GetTitle(GameEnum gameEnum)
-    {
-        return gameEnum switch
-        {
-            GameEnum.CrazyEights => CrazyEights.GetTitle(),
-            GameEnum.CrazyEightsVariation => CrazyEightsVariation.GetTitle(),
-            GameEnum.President => President.GetTitle(),
-            _ => throw new ArgumentOutOfRangeException(nameof(gameEnum), gameEnum, null)
-        };
-    }
+    public abstract string Title { get; }
 
-    public static string GetDescription(GameEnum gameEnum)
-    {
-        return gameEnum switch
-        {
-            GameEnum.CrazyEights => CrazyEights.GetDescription(),
-            GameEnum.CrazyEightsVariation => CrazyEightsVariation.GetDescription(),
-            GameEnum.President => President.GetDescription(),
-            _ => throw new ArgumentOutOfRangeException(nameof(gameEnum), gameEnum, null)
-        };
-    }
+    public abstract string Description { get; }
+
+    /// <summary>
+    /// Meaning of points
+    /// </summary>
+    /// <returns>Whether having a lot of points is good</returns>
+    public abstract bool PointsAreGood { get; }
 
     /// <summary>
     /// Setup a new game
     /// </summary>
     /// <param name="players">The amount of players in that game</param>
-    public void Initialize(int players);
+    public abstract void Initialize(int players);
 
     /// <summary>
     /// Determine the state of the game
     /// </summary>
     /// <returns>Whether the game is over</returns>
-    public bool IsOver();
+    public abstract bool IsOver();
 
     /// <summary>
     /// Calculate points
     /// </summary>
     /// <returns>The points for this round for every user</returns>
-    public List<int> CalcPoints();
+    public abstract List<int> CalcPoints();
 
     /// <summary>
     /// Gets passed to a next rounds game constructor
@@ -54,28 +42,22 @@ public interface IGameService
     }
 
     /// <summary>
-    /// Meaning of points
-    /// </summary>
-    /// <returns>Whether having a lot of points is good</returns>
-    public bool PointsAreGood();
-
-    /// <summary>
     /// Play a card
     /// </summary>
     /// <param name="id">The id of the player who tries to play</param>
     /// <param name="cardIndex">The index of the card that will be played</param>
-    public void Play(int id, int cardIndex);
+    public abstract void Play(int id, int cardIndex);
 
     /// <summary>
     /// Executes a feature, does nothing if fails
     /// </summary>
     /// <param name="id">The id of the player who tries to execute the feature</param>
     /// <param name="featureId">The id of the feature</param>
-    public void ExecuteFeature(int id, int featureId);
+    public abstract void ExecuteFeature(int id, int featureId);
 
     /// <summary>
     /// Representation of the current game, for every user
     /// </summary>
     /// <returns>A list of game data with all needed information for a player</returns>
-    public List<GameData> GetGameData();
+    public abstract List<GameData> GetGameData();
 }
