+++ src/cards/Data/Game/IGameFeature.cs
@@ -2,7 +2,8 @@ namespace cards.Data.Game;
 
 public interface IGameFeature
 {
-    public string GetName();
+    public string Name { get; }
+
     public bool IsExecutable(int player);
     public bool Execute(int player);
 }
